refactor: move sentHeadersFromTimeline to utils/timeline
